pyproject.toml
[project]
name = "pyomo-replace"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "idaes-pse==2.8.0",
    "pyomo==6.9.2",
    "ahuora_compounds@git+https://github.com/waikato-ahuora-smart-energy-systems/PropertyPackages.git@v0.0.29",
]